Get tags and author pages to work
I had to remove the github-pages plugin because it disables custom plugins.
Then I had to re-add the paginate and remote-theme dependencies.
I added placeholder layout files for the plugins to reference.

This is all a bit messy but the end result is nice. I still like this more than wordpress.